+++ packages/browser/src/browser/__tests__/typedef-tests/analytics-browser.ts
@@ -2,6 +2,7 @@ import { Analytics } from '@/core/analytics'
 import { Context } from '@/core/context'
 import { AnalyticsBrowser } from '@/browser'
 import { assertNotAny, assertIs } from '@/test-helpers/type-assertions'
+import { Group, User } from '../../../core/user'
 
 /**
  * These are general typescript definition tests;
@@ -54,4 +55,59 @@ export default {
           assertIs<number | [Analytics, Context]>(response)
         })
     },
+
+  'Group should have the correct type': () => {
+    const ajs = AnalyticsBrowser.load({ writeKey: 'foo' })
+    {
+      const grpResult = ajs.group()
+      assertIs<Promise<Group>>(grpResult)
+    }
+    {
+      const grpResult = ajs.group('foo')
+      assertIs<Promise<Context>>(grpResult)
+    }
+  },
+  'User should have the correct type': () => {
+    const ajs = AnalyticsBrowser.load({ writeKey: 'foo' })
+    {
+      const grpResult = ajs.user()
+      assertIs<Promise<User>>(grpResult)
+    }
+  },
+  'Identify should work with spread objects ': () => {
+    const user = {
+      name: 'john',
+      id: 12345,
+    }
+    const { id, ...traits } = user
+    void AnalyticsBrowser.load({ writeKey: 'foo' }).identify('foo', traits)
+  },
+  'Track should work with spread objects': () => {
+    const user = {
+      name: 'john',
+      id: 12345,
+    }
+    const { id, ...traits } = user
+    void AnalyticsBrowser.load({ writeKey: 'foo' }).track('foo', traits)
+  },
+  'Identify should work with generic objects ': () => {
+    const user = {
+      name: 'john',
+      id: 12345,
+    }
+    void AnalyticsBrowser.load({ writeKey: 'foo' }).identify('foo', user)
+  },
+  'Context should have a key allowing arbitrary properties': async () => {
+    const [_, ctx] = await AnalyticsBrowser.load({ writeKey: 'foo' })
+    const properties = ctx.event.properties!
+
+    properties.category.baz = 'hello'
+  },
+  'Track should allow undefined properties': () => {
+    type User = {
+      name?: string
+      thing: 123
+    }
+    void AnalyticsBrowser.load({ writeKey: 'foo' }).track('foo', {} as User)
+  },
 }