+++ packages/browser/src/core/events/interfaces.ts
@@ -1,10 +1,13 @@
+import { Context } from '../context'
 import { CompactMetric } from '../stats'
 import { ID } from '../user'
 
 export type JSONPrimitive = string | number | boolean | null
 export type JSONValue = JSONPrimitive | JSONObject | JSONArray
 export type JSONObject = { [member: string]: JSONValue }
-export type JSONArray = Array<JSONValue>
+export type JSONArray = JSONValue[]
+
+export type Callback = (ctx: Context) => Promise<unknown> | unknown
 
 export type Integrations = {
   All?: boolean
@@ -16,7 +19,7 @@ export type Options = {
   anonymousId?: ID
   timestamp?: Date | string
   context?: AnalyticsContext
-  traits?: object
+  traits?: Traits
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   [key: string]: any
@@ -97,6 +100,11 @@ interface AnalyticsContext {
   [key: string]: any
 }
 
+// This is not ideal, but it works with all the edge cases
+export type Traits = Record<string, any>
+
+export type EventProperties = Record<string, any>
+
 export interface SegmentEvent {
   messageId?: string
 
@@ -105,14 +113,23 @@ export interface SegmentEvent {
   // page specific
   category?: string
   name?: string
-
-  properties?: object & {
-    [k: string]: JSONValue
-  }
-
-  traits?: object & {
-    [k: string]: JSONValue
-  }
+  /**
+   * An object literal representing Segment event properties
+   * - track: https://segment.com/docs/connections/spec/track/#properties
+   * - page: https://segment.com/docs/connections/spec/page/#properties
+   * - screen: https://segment.com/docs/connections/spec/screen/#properties
+   * @example
+   *  { artistID: 2435325, songID: 13532532 }
+   */
+  properties?: EventProperties
+  /**
+   * An object literal representing traits
+   * - identify: https://segment.com/docs/connections/spec/identify/#traits
+   * - group: https://segment.com/docs/connections/spec/group/#traits
+   * @example
+   *  { name: "john", age: 25 }
+   */
+  traits?: Traits
 
   integrations?: Integrations
   context?: AnalyticsContext | Options