+++ packages/browser/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @segment/analytics-next
 
+## 1.42.3
+
+### Patch Changes
+
+- [#577](https://github.com/segmentio/analytics-next/pull/577) [`8d48bdc`](https://github.com/segmentio/analytics-next/commit/8d48bdc106a15d6368de215e11712470596e1df7) Thanks [@chrisradek](https://github.com/chrisradek)! - Fixes an issue where the 'secure' cookie setting was not being applied correctly when specified.
+
+* [#579](https://github.com/segmentio/analytics-next/pull/579) [`98504e2`](https://github.com/segmentio/analytics-next/commit/98504e2f5514030d782e06df0562ea25f4f01c34) Thanks [@silesky](https://github.com/silesky)! - Fix SegmentEvent and EventProperties and add tests
+
+## 1.42.2
+
+### Patch Changes
+
+- [#578](https://github.com/segmentio/analytics-next/pull/578) [`ab8cea3`](https://github.com/segmentio/analytics-next/commit/ab8cea3a210a161fcb7c27e07d41edcb78fa4c3f) Thanks [@silesky](https://github.com/silesky)! - Loosen SegmentEvent and Traits typescript interface (Fixes [#570](https://github.com/segmentio/analytics-next/issues/570), [#575](https://github.com/segmentio/analytics-next/issues/575)).
+
+## 1.42.1
+
+### Patch Changes
+
+- [#567](https://github.com/segmentio/analytics-next/pull/567) [`5cd9358`](https://github.com/segmentio/analytics-next/commit/5cd9358d0a8ae9711eea22ae719e9f7363581798) Thanks [@silesky](https://github.com/silesky)! - Do not allow the "user" method to change its return types over its lifecycle. We should always return a promise for wrapped methods in AnalyticsBrowser, regardless if the underlying Analytics method is sync or async.
+
+## 1.42.0
+
+### Minor Changes
+
+- [#561](https://github.com/segmentio/analytics-next/pull/561) [`1af2a9f`](https://github.com/segmentio/analytics-next/commit/1af2a9fa13596464d152bcecac19df110cd074bc) Thanks [@silesky](https://github.com/silesky)! - Add 'screen', 'register', 'deregister', 'user' method and 'VERSION' property on AnalyticsBrowser. Allow buffering of 'screen', 'register', 'deregister' methods for snippet users.
+
+### Patch Changes
+
+- Updated dependencies [[`93c8f1f`](https://github.com/segmentio/analytics-next/commit/93c8f1f7dabe6fca5bd0f8f9f0cc0c0e14cd2128)]:
+  - @segment/analytics-core@1.0.1
+
 ## 1.41.1
 
 ### Patch Changes